# app/internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateInvoiceDetailRequest:
    properties:
      ingredient_id:
        example: e2g30c86-9h7e-6647-1636-3d23effg1f1
        type: string
      invoice_id:
        example: c0f18a64-7f5c-4425-9414-1b01cddee9d9
        type: string
      price:
        example: "500000"
        type: string
      price_per_unit:
        example: "10000"
        type: string
      quantity:
        example: 50
        type: integer
    required:
    - ingredient_id
    - invoice_id
    - price
    - price_per_unit
    - quantity
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  model.InvoiceDetailResponse:
    properties:
      created_at:
        type: string
      id:
        example: d1f29b75-8g6d-5536-0525-2c12deeef0e0
        type: string
      ingredient_id:
        example: e2g30c86-9h7e-6647-1636-3d23effg1f1
        type: string
      invoice_id:
        example: c0f18a64-7f5c-4425-9414-1b01cddee9d9
        type: string
      price:
        example: "500000"
        type: string
      price_per_unit:
        example: "10000"
        type: string
      quantity:
        example: 50
        type: integer
      updated_at:
        type: string
    type: object
  model.InvoiceDetailWithIngredientResponse:
    properties:
      created_at:
        type: string
      id:
        example: d1f29b75-8g6d-5536-0525-2c12deeef0e0
        type: string
      ingredient_id:
        example: e2g30c86-9h7e-6647-1636-3d23effg1f1
        type: string
      ingredient_measurement:
        example: kg
        type: string
      ingredient_name:
        example: Tomato
        type: string
      ingredient_picture:
        example: http://example.com/tomato.jpg
        type: string
      invoice_id:
        example: c0f18a64-7f5c-4425-9414-1b01cddee9d9
        type: string
      price:
        example: "500000"
        type: string
      price_per_unit:
        example: "10000"
        type: string
      quantity:
        example: 50
        type: integer
      updated_at:
        type: string
    type: object
...
  model.UpdateInvoiceDetailQuantityRequest:
    properties:
      quantity:
        example: 50
        type: integer
    required:
    - quantity
    type: object
  model.UpdateInvoiceDetailRequest:
    properties:
      ingredient_id:
        example: e2g30c86-9h7e-6647-1636-3d23effg1f1
        type: string
      price:
        example: "500000"
        type: string
      price_per_unit:
        example: "10000"
        type: string
      quantity:
        example: 50
        type: integer
    type: object
...
host: back.staging.maryai.yurtal.tech
info:
  contact: {}
  description: MaryAI API server with multi-language support (uz, ru, en)
  title: MaryAI API
  version: "1.0"
paths:
...
  /api/v1/invoice-details:
    get:
      description: Get all invoice details with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.InvoiceDetailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all invoice details
      tags:
      - Invoice Details
    post:
      consumes:
      - application/json
      description: Create a new line item in an invoice
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Create invoice detail request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateInvoiceDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.InvoiceDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create invoice detail
      tags:
      - Invoice Details
  /api/v1/invoice-details/{id}:
    delete:
      description: Delete (soft delete) an invoice detail
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete invoice detail
      tags:
      - Invoice Details
    get:
      description: Get a single invoice detail by its ID
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InvoiceDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice detail by ID
      tags:
      - Invoice Details
    put:
      consumes:
      - application/json
      description: Update an existing invoice detail (line item)
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      - description: Update invoice detail request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateInvoiceDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InvoiceDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update invoice detail
      tags:
      - Invoice Details
  /api/v1/invoice-details/{id}/quantity:
    put:
      consumes:
      - application/json
      description: Update the quantity of a line item in an invoice
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      - description: Quantity update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateInvoiceDetailQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InvoiceDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update invoice detail quantity
      tags:
      - Invoice Details
  /api/v1/invoice-details/{id}/restore:
    post:
      description: Restore a soft-deleted invoice detail
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InvoiceDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore invoice detail
      tags:
      - Invoice Details
  /api/v1/invoice-details/{id}/with-ingredient:
    get:
      description: Get invoice detail including the associated ingredient information
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice Detail ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InvoiceDetailWithIngredientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice detail with ingredient
      tags:
      - Invoice Details
  /api/v1/invoice-details/ingredient/{ingredient_id}:
    get:
      description: Get all invoice details containing a specific ingredient with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Ingredient ID
        in: path
        name: ingredient_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.InvoiceDetailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice details by ingredient ID
      tags:
      - Invoice Details
  /api/v1/invoice-details/invoice/{invoice_id}:
    get:
      description: Get all line items for a specific invoice with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Invoice ID
        in: path
        name: invoice_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.InvoiceDetailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get invoice details by invoice ID
      tags:
      - Invoice Details
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"